--- a/app/services/service_provider_request_handler.rb
+++ b/app/services/service_provider_request_handler.rb
@@ -15,6 +15,7 @@ class ServiceProviderRequestHandler
 
     StoreSpMetadataInSession.new(session: session, request_id: request_id).call(
       service_provider_request: service_provider_request,
+      requested_service_provider: protocol.service_provider,
     )
   end