## 🛠 Summary of changes

OIDC and SAML auth are one of the hotter request paths, so we did a bit of work today to find opportunities to speed them up. This PR has two commits to implement two separate refactors. Both try to reduce duplicative work.

https://github.com/18F/identity-idp/commit/114f3d03531b60820b47044f8369e20fb4565ab6 adds a parameter to `StoreSpMetadata` to avoid querying the DB (typically ActiveRecord DB cache) for the `ServiceProvider` that has already been fetched. It also memoizes the `service_provider` method to avoid further queries in the class.

11b6cd2998e5c7e35eb7f376ce30004fc8858c2a is the bigger improvement by avoiding [apply_secure_headers_override](https://github.com/18F/identity-idp/blob/25b78167726a5aa03cd4bb2630d7f39c57cffbaa/app/controllers/concerns/secure_headers_concern.rb#L4-L11) which re-parses the request parameters, re-validates the OIDC form, and hits the database (cache) again. The change brings it closer in line with how [SAML](https://github.com/18F/identity-idp/blob/25b78167726a5aa03cd4bb2630d7f39c57cffbaa/app/controllers/saml_idp_controller.rb#L153-L158) calls `override_form_action_csp` directly with the URIs in the controller.

Together, they improve the speed of the common-case OIDC Authorization Request by ~15%. SAML benefits a little bit as a result of the first commit (~1%).

<details>
<summary>
Performance Test Results
</summary>

OIDC

```
❤️ ❤️ ❤️  (Statistically Significant) ❤️ ❤️ ❤️

(1.0330 seconds) "skip re-parsing and validating OIDC parameters"
  FASTER 🚀🚀🚀 by:
     1.1898x [older/newer]
    15.9510% [(older - newer) / older * 100]
 (1.2290 seconds) "main"

Iterations per sample: 200
Samples: 200

Test type: Kolmogorov Smirnov
Confidence level: 99.0 %
Is significant? (max > critical): true
D critical: 0.15174271293851463
D max: 0.845

Histograms (time ranges are in seconds):

   description:                                                                        description:
     "skip re-parsing and validating OIDC parameters"                                "main"
              ┌                                        ┐                ┌                                        ┐
   [0.8, 1.0) ┤▇▇▇▇▇▇▇▇▇▇▇ 43                                [0.8, 1.0) ┤ 0
   [1.0, 1.2) ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 137       [1.0, 1.2) ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 57
   [1.2, 1.4) ┤▇▇▇ 11                                        [1.2, 1.4) ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 118
   [1.4, 1.6) ┤▇ 4                                           [1.4, 1.6) ┤▇▇▇▇ 12
   [1.6, 1.8) ┤▇ 3                                           [1.6, 1.8) ┤▇▇ 6
   [1.8, 2.0) ┤ 0                                            [1.8, 2.0) ┤▇ 3
   [2.0, 2.2) ┤ 1                                            [2.0, 2.2) ┤ 0
   [2.2, 2.4) ┤ 0                                            [2.2, 2.4) ┤▇ 4
   [2.4, 2.6) ┤ 0                                            [2.4, 2.6) ┤ 0
   [2.6, 2.8) ┤ 0                                            [2.6, 2.8) ┤ 0
   [2.8, 3.0) ┤ 1                                            [2.8, 3.0) ┤ 0
              └                                        ┘                └                                        ┘
                         # of runs in range                                        # of runs in range
```

SAML

```
❤️ ❤️ ❤️  (Statistically Significant) ❤️ ❤️ ❤️

(2.4706 seconds) "skip re-parsing and validating OIDC parameters" ref: 
  FASTER 🚀🚀🚀 by:
    1.0198x [older/newer]
    1.9391% [(older - newer) / older * 100]
(2.5194 seconds) "main"

Iterations per sample: 200
Samples: 51

Test type: Kolmogorov Smirnov
Confidence level: 99.0 %
Is significant? (max > critical): true
D critical: 0.30049534876137013
D max: 0.3333333333333333

Histograms (time ranges are in seconds):

   description:                                                   description:
     "skip re-parsing and validating OIDC parameters"              "main"
              ┌                                        ┐                ┌                                        ┐
   [2.0, 2.8) ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 46       [2.0, 2.8) ┤▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 42
   [2.8, 3.6) ┤▇▇▇ 4                                         [2.8, 3.6) ┤▇▇▇▇▇ 6
   [3.5, 4.3) ┤▇ 1                                           [3.5, 4.3) ┤▇▇ 2
   [4.3, 5.1) ┤ 0                                            [4.3, 5.1) ┤ 0
   [5.0, 5.8) ┤ 0                                            [5.0, 5.8) ┤ 0
   [5.8, 6.6) ┤ 0                                            [5.8, 6.6) ┤▇ 1
              └                                        ┘                └                                        ┘
                         # of runs in range                                        # of runs in range
```